Refactor header to [MainHeader] and [AuthHeader]
