+++ src/components/navigations/Main/Header/MainHeader/Profile.svelte
@@ -14,7 +14,7 @@
 </style>
 
 <script lang="ts">
-  import CircleImage from '../../../uis/CircleImage.svelte';
+  import CircleImage from '../../../../uis/CircleImage.svelte';
 
   export let name: string;
   export let imageSrc: string;
