+++ src/components/navigations/Main/Header/MainHeader/index.svelte
@@ -23,8 +23,8 @@
 
 <script lang="ts">
   import {push, replace} from 'svelte-spa-router';
-  import {user} from '../../../../stores/sessionStore';
-  import {SvgBell} from '../../../../utils/Icon';
+  import {user} from '../../../../../stores/sessionStore';
+  import {SvgBell} from '../../../../../utils/Icon';
   import Profile from './Profile.svelte';
   import Search from './Serach.svelte';
