+++ src/components/navigations/Main/Header/index.svelte
@@ -28,12 +28,10 @@
 </style>
 
 <script lang="ts">
-  import {_} from 'svelte-i18n';
-  import {push} from 'svelte-spa-router';
   import {user} from '../../../../stores/sessionStore';
   import {SvgLogo} from '../../../../utils/Icon';
-  import Button from '../../../uis/Button.svelte';
-  import Manipulation from './Manipulation.svelte';
+  import AuthHeader from './AuthHeader.svelte';
+  import MainHeader from './MainHeader/index.svelte';
 </script>
 
 <nav>
@@ -41,25 +39,9 @@
     <SvgLogo width="34.29" />
     <div class="nav-menu">
       {#if $user}
-        <Manipulation />
+        <MainHeader />
       {:else}
-        <Button
-          on:click={async () => {
-            await push('/sign_in');
-          }}
-          style="color: black; font-size: 14px; align-self: stretch;"
-        >
-          <div class="text">{$_('login')}</div>
-        </Button>
-        <Button
-          on:click={() => {
-            // eslint-disable-next-line @typescript-eslint/no-floating-promises
-            push('/sign_up');
-          }}
-          style="font-size: 14px; margin-left: 12px; align-self: stretch;"
-        >
-          <div class="text">{$_('SignUp.title')}</div>
-        </Button>
+        <AuthHeader />
       {/if}
     </div>
   </div>
